better indexing of pydantic models
