+++ tests/contrib/test_pydantic.py
@@ -67,7 +67,7 @@ class TestPydantic(test.TestCase):
             self.Event_Pydantic.model_json_schema(),
             {
                 "$defs": {
-                    "Address_diegkp_leaf": {
+                    "Address_4smkxs_leaf": {
                         "additionalProperties": False,
                         "properties": {
                             "city": {"maxLength": 64, "title": "City", "type": "string"},
@@ -83,7 +83,7 @@ class TestPydantic(test.TestCase):
                         "title": "Address",
                         "type": "object",
                     },
-                    "Reporter_flpkb7_leaf": {
+                    "Reporter_4tvqui_leaf": {
                         "additionalProperties": False,
                         "description": "Whom is assigned as the reporter",
                         "properties": {
@@ -99,7 +99,7 @@ class TestPydantic(test.TestCase):
                         "title": "Reporter",
                         "type": "object",
                     },
-                    "Team_uqjfiz_leaf": {
+                    "Team_bnxjyv_leaf": {
                         "additionalProperties": False,
                         "description": "Team that is a playing",
                         "properties": {
@@ -127,7 +127,7 @@ class TestPydantic(test.TestCase):
                         "title": "Team",
                         "type": "object",
                     },
-                    "Tournament_lloz5q_leaf": {
+                    "Tournament_djsht2_leaf": {
                         "additionalProperties": False,
                         "properties": {
                             "id": {
@@ -165,13 +165,13 @@ class TestPydantic(test.TestCase):
                     },
                     "name": {"description": "The name", "title": "Name", "type": "string"},
                     "tournament": {
-                        "$ref": "#/$defs/Tournament_lloz5q_leaf",
+                        "$ref": "#/$defs/Tournament_djsht2_leaf",
                         "description": "What tournaments is a happenin'",
                     },
                     "reporter": {
                         "anyOf": [
                             {
-                                "$ref": "#/$defs/Reporter_flpkb7_leaf"
+                                "$ref": "#/$defs/Reporter_4tvqui_leaf"
                             },
                             {"type": "null"},
                         ],
@@ -180,7 +180,7 @@ class TestPydantic(test.TestCase):
                     },
                     "participants": {
                         "items": {
-                            "$ref": "#/$defs/Team_uqjfiz_leaf"
+                            "$ref": "#/$defs/Team_bnxjyv_leaf"
                         },
                         "title": "Participants",
                         "type": "array",
@@ -203,7 +203,7 @@ class TestPydantic(test.TestCase):
                     "address": {
                         "anyOf": [
                             {
-                                "$ref": "#/$defs/Address_diegkp_leaf"
+                                "$ref": "#/$defs/Address_4smkxs_leaf"
                             },
                             {"type": "null"},
                         ],
@@ -232,7 +232,7 @@ class TestPydantic(test.TestCase):
             self.Event_Pydantic_List.model_json_schema(),
             {
                 "$defs": {
-                    "Event_vh77nq": {
+                    "Event_c5pgat": {
                         "additionalProperties": False,
                         "description": "Events on the calendar",
                         "properties": {
@@ -244,13 +244,13 @@ class TestPydantic(test.TestCase):
                             },
                             "name": {"description": "The name", "title": "Name", "type": "string"},
                             "tournament": {
-                                "$ref": "#/$defs/Tournament_lloz5q_leaf",
+                                "$ref": "#/$defs/Tournament_djsht2_leaf",
                                 "description": "What tournaments is a happenin'",
                             },
                             "reporter": {
                                 "anyOf": [
                                     {
-                                        "$ref": "#/$defs/Reporter_flpkb7_leaf"
+                                        "$ref": "#/$defs/Reporter_4tvqui_leaf"
                                     },
                                     {"type": "null"},
                                 ],
@@ -259,7 +259,7 @@ class TestPydantic(test.TestCase):
                             },
                             "participants": {
                                 "items": {
-                                    "$ref": "#/$defs/Team_uqjfiz_leaf"
+                                    "$ref": "#/$defs/Team_bnxjyv_leaf"
                                 },
                                 "title": "Participants",
                                 "type": "array",
@@ -289,7 +289,7 @@ class TestPydantic(test.TestCase):
                             "address": {
                                 "anyOf": [
                                     {
-                                        "$ref": "#/$defs/Address_diegkp_leaf"
+                                        "$ref": "#/$defs/Address_4smkxs_leaf"
                                     },
                                     {"type": "null"},
                                 ],
@@ -311,7 +311,7 @@ class TestPydantic(test.TestCase):
                         "title": "Event",
                         "type": "object",
                     },
-                    "Address_diegkp_leaf": {
+                    "Address_4smkxs_leaf": {
                         "additionalProperties": False,
                         "properties": {
                             "city": {"maxLength": 64, "title": "City", "type": "string"},
@@ -327,7 +327,7 @@ class TestPydantic(test.TestCase):
                         "title": "Address",
                         "type": "object",
                     },
-                    "Reporter_flpkb7_leaf": {
+                    "Reporter_4tvqui_leaf": {
                         "additionalProperties": False,
                         "description": "Whom is assigned as the reporter",
                         "properties": {
@@ -343,7 +343,7 @@ class TestPydantic(test.TestCase):
                         "title": "Reporter",
                         "type": "object",
                     },
-                    "Team_uqjfiz_leaf": {
+                    "Team_bnxjyv_leaf": {
                         "additionalProperties": False,
                         "description": "Team that is a playing",
                         "properties": {
@@ -371,7 +371,7 @@ class TestPydantic(test.TestCase):
                         "title": "Team",
                         "type": "object",
                     },
-                    "Tournament_lloz5q_leaf": {
+                    "Tournament_djsht2_leaf": {
                         "additionalProperties": False,
                         "properties": {
                             "id": {
@@ -399,7 +399,7 @@ class TestPydantic(test.TestCase):
                     },
                 },
                 "description": "Events on the calendar",
-                "items": {"$ref": "#/$defs/Event_vh77nq"},
+                "items": {"$ref": "#/$defs/Event_c5pgat"},
                 "title": "Event_list",
                 "type": "array",
             },
@@ -410,7 +410,7 @@ class TestPydantic(test.TestCase):
             self.Address_Pydantic.model_json_schema(),
             {
                 "$defs": {
-                    "Event_az5m74_leaf": {
+                    "Event_jz7ivr_leaf": {
                         "additionalProperties": False,
                         "description": "Events on the calendar",
                         "properties": {
@@ -422,13 +422,13 @@ class TestPydantic(test.TestCase):
                             },
                             "name": {"description": "The name", "title": "Name", "type": "string"},
                             "tournament": {
-                                "$ref": "#/$defs/Tournament_lloz5q_leaf",
+                                "$ref": "#/$defs/Tournament_djsht2_leaf",
                                 "description": "What tournaments is a happenin'",
                             },
                             "reporter": {
                                 "anyOf": [
                                     {
-                                        "$ref": "#/$defs/Reporter_flpkb7_leaf"
+                                        "$ref": "#/$defs/Reporter_4tvqui_leaf"
                                     },
                                     {"type": "null"},
                                 ],
@@ -437,7 +437,7 @@ class TestPydantic(test.TestCase):
                             },
                             "participants": {
                                 "items": {
-                                    "$ref": "#/$defs/Team_uqjfiz_leaf"
+                                    "$ref": "#/$defs/Team_bnxjyv_leaf"
                                 },
                                 "title": "Participants",
                                 "type": "array",
@@ -478,7 +478,7 @@ class TestPydantic(test.TestCase):
                         "title": "Event",
                         "type": "object",
                     },
-                    "Reporter_flpkb7_leaf": {
+                    "Reporter_4tvqui_leaf": {
                         "additionalProperties": False,
                         "description": "Whom is assigned as the reporter",
                         "properties": {
@@ -494,7 +494,7 @@ class TestPydantic(test.TestCase):
                         "title": "Reporter",
                         "type": "object",
                     },
-                    "Team_uqjfiz_leaf": {
+                    "Team_bnxjyv_leaf": {
                         "additionalProperties": False,
                         "description": "Team that is a playing",
                         "properties": {
@@ -522,7 +522,7 @@ class TestPydantic(test.TestCase):
                         "title": "Team",
                         "type": "object",
                     },
-                    "Tournament_lloz5q_leaf": {
+                    "Tournament_djsht2_leaf": {
                         "additionalProperties": False,
                         "properties": {
                             "id": {
@@ -553,7 +553,7 @@ class TestPydantic(test.TestCase):
                 "properties": {
                     "city": {"maxLength": 64, "title": "City", "type": "string"},
                     "street": {"maxLength": 128, "title": "Street", "type": "string"},
-                    "event": {"$ref": "#/$defs/Event_az5m74_leaf"},
+                    "event": {"$ref": "#/$defs/Event_jz7ivr_leaf"},
                     "event_id": {
                         "maximum": 9223372036854775807,
                         "minimum": -9223372036854775808,
@@ -572,7 +572,7 @@ class TestPydantic(test.TestCase):
             self.Tournament_Pydantic.model_json_schema(),
             {
                 "$defs": {
-                    "Event_atwjj7_leaf": {
+                    "Event_yu5egy_leaf": {
                         "additionalProperties": False,
                         "description": "Events on the calendar",
                         "properties": {
@@ -586,7 +586,7 @@ class TestPydantic(test.TestCase):
                             "reporter": {
                                 "anyOf": [
                                     {
-                                        "$ref": "#/$defs/Reporter_flpkb7_leaf"
+                                        "$ref": "#/$defs/Reporter_4tvqui_leaf"
                                     },
                                     {"type": "null"},
                                 ],
@@ -595,7 +595,7 @@ class TestPydantic(test.TestCase):
                             },
                             "participants": {
                                 "items": {
-                                    "$ref": "#/$defs/Team_uqjfiz_leaf"
+                                    "$ref": "#/$defs/Team_bnxjyv_leaf"
                                 },
                                 "title": "Participants",
                                 "type": "array",
@@ -625,7 +625,7 @@ class TestPydantic(test.TestCase):
                             "address": {
                                 "anyOf": [
                                     {
-                                        "$ref": "#/$defs/Address_diegkp_leaf"
+                                        "$ref": "#/$defs/Address_4smkxs_leaf"
                                     },
                                     {"type": "null"},
                                 ],
@@ -646,7 +646,7 @@ class TestPydantic(test.TestCase):
                         "title": "Event",
                         "type": "object",
                     },
-                    "Address_diegkp_leaf": {
+                    "Address_4smkxs_leaf": {
                         "additionalProperties": False,
                         "properties": {
                             "city": {"maxLength": 64, "title": "City", "type": "string"},
@@ -662,7 +662,7 @@ class TestPydantic(test.TestCase):
                         "title": "Address",
                         "type": "object",
                     },
-                    "Reporter_flpkb7_leaf": {
+                    "Reporter_4tvqui_leaf": {
                         "additionalProperties": False,
                         "description": "Whom is assigned as the reporter",
                         "properties": {
@@ -678,7 +678,7 @@ class TestPydantic(test.TestCase):
                         "title": "Reporter",
                         "type": "object",
                     },
-                    "Team_uqjfiz_leaf": {
+                    "Team_bnxjyv_leaf": {
                         "additionalProperties": False,
                         "description": "Team that is a playing",
                         "properties": {
@@ -724,7 +724,7 @@ class TestPydantic(test.TestCase):
                     },
                     "events": {
                         "description": "What tournaments is a happenin'",
-                        "items": {"$ref": "#/$defs/Event_atwjj7_leaf"},
+                        "items": {"$ref": "#/$defs/Event_yu5egy_leaf"},
                         "title": "Events",
                         "type": "array",
                     },
@@ -740,7 +740,7 @@ class TestPydantic(test.TestCase):
             self.Team_Pydantic.model_json_schema(),
             {
                 "$defs": {
-                    "Event_37gjqu_leaf": {
+                    "Event_7fv6fv_leaf": {
                         "additionalProperties": False,
                         "description": "Events on the calendar",
                         "properties": {
@@ -752,13 +752,13 @@ class TestPydantic(test.TestCase):
                             },
                             "name": {"description": "The name", "title": "Name", "type": "string"},
                             "tournament": {
-                                "$ref": "#/$defs/Tournament_lloz5q_leaf",
+                                "$ref": "#/$defs/Tournament_djsht2_leaf",
                                 "description": "What tournaments is a happenin'",
                             },
                             "reporter": {
                                 "anyOf": [
                                     {
-                                        "$ref": "#/$defs/Reporter_flpkb7_leaf"
+                                        "$ref": "#/$defs/Reporter_4tvqui_leaf"
                                     },
                                     {"type": "null"},
                                 ],
@@ -790,7 +790,7 @@ class TestPydantic(test.TestCase):
                             "address": {
                                 "anyOf": [
                                     {
-                                        "$ref": "#/$defs/Address_diegkp_leaf"
+                                        "$ref": "#/$defs/Address_4smkxs_leaf"
                                     },
                                     {"type": "null"},
                                 ],
@@ -811,7 +811,7 @@ class TestPydantic(test.TestCase):
                         "title": "Event",
                         "type": "object",
                     },
-                    "Address_diegkp_leaf": {
+                    "Address_4smkxs_leaf": {
                         "additionalProperties": False,
                         "properties": {
                             "city": {"maxLength": 64, "title": "City", "type": "string"},
@@ -827,7 +827,7 @@ class TestPydantic(test.TestCase):
                         "title": "Address",
                         "type": "object",
                     },
-                    "Reporter_flpkb7_leaf": {
+                    "Reporter_4tvqui_leaf": {
                         "additionalProperties": False,
                         "description": "Whom is assigned as the reporter",
                         "properties": {
@@ -843,7 +843,7 @@ class TestPydantic(test.TestCase):
                         "title": "Reporter",
                         "type": "object",
                     },
-                    "Tournament_lloz5q_leaf": {
+                    "Tournament_djsht2_leaf": {
                         "additionalProperties": False,
                         "properties": {
                             "id": {
@@ -889,7 +889,7 @@ class TestPydantic(test.TestCase):
                         "title": "Alias",
                     },
                     "events": {
-                        "items": {"$ref": "#/$defs/Event_37gjqu_leaf"},
+                        "items": {"$ref": "#/$defs/Event_7fv6fv_leaf"},
                         "title": "Events",
                         "type": "array",
                     },
@@ -1293,7 +1293,7 @@ class TestPydanticCycle(test.TestCase):
             self.Employee_Pydantic.model_json_schema(),
             {
                 "$defs": {
-                    "Employee_dkvdqq_leaf": {
+                    "Employee_lqnwvu_leaf": {
                         "additionalProperties": False,
                         "properties": {
                             "id": {
@@ -1304,7 +1304,7 @@ class TestPydanticCycle(test.TestCase):
                             },
                             "name": {"maxLength": 50, "title": "Name", "type": "string"},
                             "talks_to": {
-                                "items": {"$ref": "#/$defs/Employee_pb36jg_leaf"},
+                                "items": {"$ref": "#/$defs/Employee_okqh4k_leaf"},
                                 "title": "Talks To",
                                 "type": "array",
                             },
@@ -1321,7 +1321,7 @@ class TestPydanticCycle(test.TestCase):
                                 "title": "Manager Id",
                             },
                             "team_members": {
-                                "items": {"$ref": "#/$defs/Employee_pb36jg_leaf"},
+                                "items": {"$ref": "#/$defs/Employee_okqh4k_leaf"},
                                 "title": "Team Members",
                                 "type": "array",
                             },
@@ -1330,7 +1330,7 @@ class TestPydanticCycle(test.TestCase):
                         "title": "Employee",
                         "type": "object",
                     },
-                    "Employee_pb36jg_leaf": {
+                    "Employee_okqh4k_leaf": {
                         "additionalProperties": False,
                         "properties": {
                             "id": {
@@ -1368,7 +1368,7 @@ class TestPydanticCycle(test.TestCase):
                     },
                     "name": {"maxLength": 50, "title": "Name", "type": "string"},
                     "talks_to": {
-                        "items": {"$ref": "#/$defs/Employee_dkvdqq_leaf"},
+                        "items": {"$ref": "#/$defs/Employee_lqnwvu_leaf"},
                         "title": "Talks To",
                         "type": "array",
                     },
@@ -1381,7 +1381,7 @@ class TestPydanticCycle(test.TestCase):
                         "title": "Manager Id",
                     },
                     "team_members": {
-                        "items": {"$ref": "#/$defs/Employee_dkvdqq_leaf"},
+                        "items": {"$ref": "#/$defs/Employee_lqnwvu_leaf"},
                         "title": "Team Members",
                         "type": "array",
                     },
@@ -1491,7 +1491,58 @@ class TestPydanticComputed(test.TestCase):
             self.Employee_Pydantic.model_json_schema(mode="serialization"),
             {
                 "$defs": {
-                    "Employee_dkvdqq_leaf": {
+                    "Employee_okqh4k_leaf": {
+                        "additionalProperties": False,
+                        "properties": {
+                            "id": {
+                                "maximum": 2147483647,
+                                "minimum": -2147483648,
+                                "title": "Id",
+                                "type": "integer"
+                            },
+                            "name": {
+                                "maxLength": 50,
+                                "title": "Name",
+                                "type": "string"
+                            },
+                            "manager_id": {
+                                "anyOf": [
+                                    {
+                                        "maximum": 2147483647,
+                                        "minimum": -2147483648,
+                                        "type": "integer"
+                                    },
+                                    {
+                                        "type": "null"
+                                    }
+                                ],
+                                "nullable": True,
+                                "title": "Manager Id"
+                            },
+                            "name_length": {
+                                "description": "",
+                                "readOnly": True,
+                                "title": "Name Length",
+                                "type": "integer"
+                            },
+                            "team_size": {
+                                "description": "Computes team size.<br/><br/>Note that this function needs to be annotated with a return type so that pydantic can<br/> generate a valid schema.<br/><br/>Note that the pydantic serializer can't call async methods, but the tortoise helpers<br/> pre-fetch relational data, so that it is available before serialization. So we don't<br/> need to await the relation. We do however have to protect against the case where no<br/> prefetching was done, hence catching and handling the<br/> ``tortoise.exceptions.NoValuesFetched`` exception.",
+                                "readOnly": True,
+                                "title": "Team Size",
+                                "type": "integer"
+                            }
+                        },
+                        "required": [
+                            "id",
+                            "name",
+                            "manager_id",
+                            "name_length",
+                            "team_size"
+                        ],
+                        "title": "Employee",
+                        "type": "object"
+                    },
+                    "Employee_lqnwvu_leaf": {
                         "additionalProperties": False,
                         "properties": {
                             "id": {
@@ -1507,7 +1558,7 @@ class TestPydanticComputed(test.TestCase):
                             },
                             "talks_to": {
                                 "items": {
-                                    "$ref": "#/$defs/Employee_pb36jg_leaf"
+                                    "$ref": "#/$defs/Employee_okqh4k_leaf"
                                 },
                                 "title": "Talks To",
                                 "type": "array"
@@ -1528,7 +1579,7 @@ class TestPydanticComputed(test.TestCase):
                             },
                             "team_members": {
                                 "items": {
-                                    "$ref": "#/$defs/Employee_pb36jg_leaf"
+                                    "$ref": "#/$defs/Employee_okqh4k_leaf"
                                 },
                                 "title": "Team Members",
                                 "type": "array"
@@ -1557,57 +1608,6 @@ class TestPydanticComputed(test.TestCase):
                         ],
                         "title": "Employee",
                         "type": "object"
-                    },
-                    "Employee_pb36jg_leaf": {
-                        "additionalProperties": False,
-                        "properties": {
-                            "id": {
-                                "maximum": 2147483647,
-                                "minimum": -2147483648,
-                                "title": "Id",
-                                "type": "integer"
-                            },
-                            "name": {
-                                "maxLength": 50,
-                                "title": "Name",
-                                "type": "string"
-                            },
-                            "manager_id": {
-                                "anyOf": [
-                                    {
-                                        "maximum": 2147483647,
-                                        "minimum": -2147483648,
-                                        "type": "integer"
-                                    },
-                                    {
-                                        "type": "null"
-                                    }
-                                ],
-                                "nullable": True,
-                                "title": "Manager Id"
-                            },
-                            "name_length": {
-                                "description": "",
-                                "readOnly": True,
-                                "title": "Name Length",
-                                "type": "integer"
-                            },
-                            "team_size": {
-                                "description": "Computes team size.<br/><br/>Note that this function needs to be annotated with a return type so that pydantic can<br/> generate a valid schema.<br/><br/>Note that the pydantic serializer can't call async methods, but the tortoise helpers<br/> pre-fetch relational data, so that it is available before serialization. So we don't<br/> need to await the relation. We do however have to protect against the case where no<br/> prefetching was done, hence catching and handling the<br/> ``tortoise.exceptions.NoValuesFetched`` exception.",
-                                "readOnly": True,
-                                "title": "Team Size",
-                                "type": "integer"
-                            }
-                        },
-                        "required": [
-                            "id",
-                            "name",
-                            "manager_id",
-                            "name_length",
-                            "team_size"
-                        ],
-                        "title": "Employee",
-                        "type": "object"
                     }
                 },
                 "additionalProperties": False,
@@ -1625,7 +1625,7 @@ class TestPydanticComputed(test.TestCase):
                     },
                     "talks_to": {
                         "items": {
-                            "$ref": "#/$defs/Employee_dkvdqq_leaf"
+                            "$ref": "#/$defs/Employee_lqnwvu_leaf"
                         },
                         "title": "Talks To",
                         "type": "array"
@@ -1646,7 +1646,7 @@ class TestPydanticComputed(test.TestCase):
                     },
                     "team_members": {
                         "items": {
-                            "$ref": "#/$defs/Employee_dkvdqq_leaf"
+                            "$ref": "#/$defs/Employee_lqnwvu_leaf"
                         },
                         "title": "Team Members",
                         "type": "array"
